+++ inc/customizer/css-output/css.php
@@ -41,5 +41,5 @@ if ( OCEANWP_LEARNDASH_ACTIVE ) {
 	require OCEANWP_INC_DIR . 'customizer/css-output/learndash.php';
 }
 
-// If Accessibility CSS is activated.
-// require OCEANWP_INC_DIR . 'customizer/css-output/a11y.php';
+// Accessibility inline CSS output.
+require OCEANWP_INC_DIR . 'customizer/css-output/a11y.php';