+++ partials/header/style/medium-header-search.php
@@ -14,7 +14,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 $search_post_type = get_theme_mod( 'ocean_menu_search_source', 'any' );
 
 $display_label     = oceanwp_is_accessibility_feature_enabled( 'ocean_display_header_search_form_label' );
-$custom_label_text = get_theme_mod( 'ocean_custom_header_search_form_label', __( 'Search this website', 'oceanwp' ) );
+$custom_label_text = oceanwp_get_header_search_form_label();
 
 // Fallback to layout default theme string if custom input text is empty.
 $default_label_text = oceanwp_theme_strings( 'owp-string-mobile-search', false );